Make ProcessingDetails view more generic
It can now accept both Evidence and Payment objects.
